+++ src/App.tsx
@@ -1,45 +1,32 @@
-import { useEffect, useState } from 'react';
-import SearchSection from './components/SearchSection/SearchSection';
-import CardSection from './components/CardSection/CardSection';
 import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary';
-import ErrorButton from './components/ErrorButton/ErrorButton';
-import Loader from './components/Loader/Loader';
-import { fetchCharacters } from './api/api';
-import { type Character } from './ts/interfaces/interfaces';
+import { Routes, Route } from 'react-router-dom';
+import CharactersPage from './pages/CharactersPage/CharactersPage';
+import HomePage from './pages/HomePage/HomePage';
 import './App.css';
 
 function App() {
-  const [characters, setCharacters] = useState<Character[]>([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const fetchData = async (searchValue: string | undefined) => {
-    try {
-      setLoading(true);
-      setError(null);
-      const data = await fetchCharacters(searchValue);
-      setCharacters(data.results);
-    } catch (error: unknown) {
-      const message = error instanceof Error ? error.message : 'Unknown error';
-      setCharacters([]);
-      setError(message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    const savedSearch = localStorage.getItem('searchValue') || '';
-    fetchData(savedSearch);
-  }, []);
+  const routes = [
+    {
+      path: '*',
+      element: <h1>404 - Page Not Found</h1>,
+    },
+    {
+      path: '/otabek996-REACT2025Q3',
+      element: <HomePage />,
+    },
+    {
+      path: '/otabek996-REACT2025Q3/characters',
+      element: <CharactersPage />,
+    },
+  ];
 
   return (
     <ErrorBoundary>
-      <SearchSection fetchData={fetchData} />
-      {loading && <Loader />}
-      {error && <div className="text-red-500 text-center">{error}</div>}
-      <CardSection characters={characters} />
-      <ErrorButton />
+      <Routes>
+        {routes.map((route, index) => (
+          <Route key={index} path={route.path} element={route.element} />
+        ))}
+      </Routes>
     </ErrorBoundary>
   );
 }
