+++ src/pages/HomePage/HomePage.tsx
@@ -1,0 +1,5 @@
+function HomePage() {
+  return <h1>This is HOME PAGE</h1>;
+}
+
+export default HomePage;
